% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ais_gather.R
\name{ais_gather}
\alias{ais_gather}
\title{Gather all AIS data}
\usage{
ais_gather(con, VID, trim.harbour = TRUE)
}
\arguments{
\item{con}{oracle connection}

\item{VID}{Vessel id}

\item{trim.harbour}{exclude points in harbour (default TRUE)}
}
\value{
a tibble
}
\description{
Gather all AIS data
}
