% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biota.R
\name{bi_len}
\alias{bi_len}
\title{bi_len}
\usage{
bi_len(con, std = TRUE, trim = TRUE)
}
\arguments{
\item{con}{oracle connection}

\item{std}{create standardized/shortcut names (default is TRUE)}

\item{trim}{trim return only key variables (default is TRUE). only operational
if std is TRUE}
}
\value{
a tibble query
}
\description{
Reads biota.lengd_skalad_v that includes a raising factor and
provides weight as a function of length. No raising done.
}
\details{
NOTE: If species only counted at station, not included in algorithm.
}
