% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biota.R
\name{bi_lwcoeffs}
\alias{bi_lwcoeffs}
\title{Length weight coefficent of some selected species}
\usage{
bi_lwcoeffs(con)
}
\arguments{
\item{con}{oracle connection}
}
\value{
A query with sid (species number), a and b
}
\description{
Length weight coefficent of some selected species
}
