% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biota.R
\name{bi_measure}
\alias{bi_measure}
\title{Biological measuremnts}
\usage{
bi_measure(con, std = TRUE, trim = TRUE)
}
\arguments{
\item{con}{oracle connection}

\item{std}{create standardized/shortcut names (default is TRUE)}

\item{trim}{trim return only key variables (default is TRUE)}
}
\value{
a query
}
\description{
Biological measuremnts
}
