% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel.R
\name{ch_sample_class}
\alias{ch_sample_class}
\title{Sample class}
\usage{
ch_sample_class(con, std = TRUE, trim = TRUE)
}
\arguments{
\item{con}{oracle connection}

\item{std}{create standardized/shortcut names (default is TRUE)}

\item{trim}{trim return only key variables (default is TRUE). only operational
if std is TRUE}
}
\value{
a tibble query
}
\description{
Reads channel.sample_category
}
