% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel.R
\name{ch_station}
\alias{ch_station}
\title{ch_station}
\usage{
ch_station(con, std = TRUE, trim = TRUE)
}
\arguments{
\item{con}{oracle connection}

\item{std}{boolean (default TRUE), create standardized/shortcut names}

\item{trim}{boolean (default TRUE), return only key variables. only operational
if std is TRUE}
}
\value{
a tibble query
}
\description{
Reads channel.stod_v. Access the same view as mar::les_stod
}
