% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{con_mar}
\alias{con_mar}
\title{Connect to MFRI database}
\usage{
con_mar(...)
}
\arguments{
\item{...}{Additional arguments to DBI functions}
}
\value{
A connection to the Oracle database
}
\description{
Connect to MFRI database
}
