% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{connect_mar}
\alias{connect_mar}
\title{Connection}
\usage{
connect_mar(dbname = "sjor", ...)
}
\arguments{
\item{dbname}{database name (default is sjor)}

\item{...}{additional argument to pass to dbConnect}
}
\value{
a connection
}
\description{
Connection
}
