% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gear.R
\name{gear_isscfg}
\alias{gear_isscfg}
\title{Title}
\usage{
gear_isscfg(con)

gear_isscfg(con)
}
\arguments{
\item{con}{}
}
\value{
A query


}
\description{
International Standard Statistical Classification of Fishing Gear (ISSCFG), see http://www.fao.org/fishery/docs/DOCUMENT/cwp/handbook/annex/AnnexM2fishinggear.pdf
}
