% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gid.R
\name{gid_correction}
\alias{gid_correction}
\title{Gear correction table}
\usage{
gid_correction(con)
}
\arguments{
\item{con}{oracle connection}
}
\value{
a query containing: visir, id and gid where the latter is the corrected
gear identification
}
\description{
Gear correction table
}
