% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gid.R
\name{gid_isscfg}
\alias{gid_isscfg}
\title{ISSCFG codes}
\usage{
gid_isscfg(con)
}
\arguments{
\item{con}{oracle connection}
}
\value{
A query
}
\description{
International Standard Statistical Classification of Fishing Gear (ISSCFG),
see http://www.fao.org/fishery/docs/DOCUMENT/cwp/handbook/annex/AnnexM2fishinggear.pdf and
https://www.fao.org/documents/card/en/c/cb4966en
}
