% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gid.R
\name{gid_isscfg_category}
\alias{gid_isscfg_category}
\title{ISSCFG category table}
\usage{
gid_isscfg_category(con)
}
\arguments{
\item{con}{oracle connection}
}
\value{
A query
}
\description{
International Standard Statistical Classification of Fishing Gear (ISSCFG), see http://www.fao.org/fishery/docs/DOCUMENT/cwp/handbook/annex/AnnexM2fishinggear.pdf
}
