% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb.R
\name{lb_base}
\alias{lb_base}
\title{Logbook base (stofn)}
\usage{
lb_base(con, correct_gear = FALSE)
}
\arguments{
\item{con}{oracle connection}

\item{correct_gear}{a boolean (default FALSE) checks for lookup-table for
gear correction (adds variable "gidc" to the tibble)}
}
\value{
sql query and data
}
\description{
Logbook base (stofn)
}
