% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_new.R
\name{lb_catch_new}
\alias{lb_catch_new}
\title{Logbook catch by species}
\usage{
lb_catch_new(con)
}
\arguments{
\item{con}{oracle connection}
}
\value{
sql query and data
}
\description{
Logbook catch by species
}
