% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_new.R
\name{lb_mobile_new}
\alias{lb_mobile_new}
\title{Logbook mobile (active) gear}
\usage{
lb_mobile_new(con, standaridize = TRUE)
}
\arguments{
\item{con}{oracle connection}

\item{correct_gear}{a boolean (default TRUE) checks for lookup-table for
gear correction (adds variable "gidc" to the tibble)}

\item{trim}{trim variables returned (default TRUE)}
}
\value{
a sql tibble
}
\description{
Logbook mobile (active) gear
}
