% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb.R
\name{lb_seine}
\alias{lb_seine}
\title{Logbook pelagic seines}
\usage{
lb_seine(con, correct_gear = TRUE, trim = TRUE)
}
\arguments{
\item{con}{Oracle connection}

\item{correct_gear}{a boolean (default TRUE) checks for lookup-table for
gear correction (adds variable "gidc" to the tibble)}

\item{trim}{trim variables returned (default TRUE)}
}
\value{
A sql tibble
}
\description{
Logbook pelagic seines
}
