% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb.R
\name{lb_static}
\alias{lb_static}
\title{Logbook static (passsive) gear}
\usage{
lb_static(con, correct_gear = FALSE)
}
\arguments{
\item{con}{oracle connection}

\item{correct_gear}{a boolean (default FALSE) checks for lookup-table for
gear correction (adds variable "gidc" to the tibble)}
}
\value{
A sql tibble
}
\description{
Logbook static (passsive) gear
}
