% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_new.R
\name{lb_static_new}
\alias{lb_static_new}
\title{Logbook new - trawl and seine}
\usage{
lb_static_new(con, trim = TRUE)
}
\arguments{
\item{con}{oracle connection}

\item{trim}{trim variables returned (default TRUE)}
}
\value{
a sql tibble
}
\description{
Logbook new - trawl and seine
}
