% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb.R
\name{lb_trail}
\alias{lb_trail}
\title{Logbook trails}
\usage{
lb_trail(con, std = TRUE, trim = TRUE)
}
\arguments{
\item{con}{Oracle connection}

\item{std}{Boolean (default TRUE)}

\item{trim}{Boolean (default TRUE)}
}
\value{
A query
}
\description{
Logbook trails
}
