% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ln.R
\name{ln_1966_1981}
\alias{ln_1966_1981}
\title{landed catch (fiskifelagid.vigtarskra66_81)}
\usage{
ln_1966_1981(con, standardize = TRUE)
}
\arguments{
\item{con}{oracle connection}

\item{standardize}{Standardize (default TRUE) renames some field and then
some but returns all variables}
}
\value{
a query
}
\description{
landed catch (fiskifelagid.vigtarskra66_81)
}
