% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ln.R
\name{ln_1982_1993}
\alias{ln_1982_1993}
\title{landed catch (fiskifelagid.landed_catch_pre94)}
\usage{
ln_1982_1993(con, standardize = TRUE)
}
\arguments{
\item{con}{oracle connection}

\item{standardize}{Standardize (default TRUE) renames some field and then
some but returns all variables}
}
\value{
a query
}
\description{
landed catch (fiskifelagid.landed_catch_pre94)
}
