% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ln.R
\name{ln_agf}
\alias{ln_agf}
\title{Landings (gafl)}
\usage{
ln_agf(con, trim = TRUE)
}
\arguments{
\item{con}{oracle connection}

\item{trim}{only return essential variables (default)}
}
\description{
Landings (gafl)
}
