% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ln.R
\name{ln_catch}
\alias{ln_catch}
\title{landed catch (kvoti.lods_oslaegt)}
\usage{
ln_catch(con, standardize = TRUE)
}
\arguments{
\item{con}{oracle connection}

\item{standardize}{Standardize (default TRUE) renames some field and then
some but returns all variables}
}
\description{
landed catch (kvoti.lods_oslaegt)
}
