% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ln.R
\name{ln_landing}
\alias{ln_landing}
\title{landings (kvoti.lods_londun)}
\usage{
ln_landing(con, standardize = TRUE)
}
\arguments{
\item{con}{oracle connection}

\item{standardize}{Standardize (default TRUE) renames some field and then
some but returns all variables}
}
\description{
landings (kvoti.lods_londun)
}
