% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{mar_fields}
\alias{mar_fields}
\title{mar_fields}
\usage{
mar_fields(mar, table)
}
\arguments{
\item{mar}{src_oracle tenging við oracle}

\item{table}{character vector specifying table names, e.g. "channel.stod_v"}
}
\value{
dataframe
}
\description{
Fallið ...
}
