% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmsi.R
\name{mmsi_icelandic_registry}
\alias{mmsi_icelandic_registry}
\title{MMSI Icelandic registry}
\source{
\url{https://www.fjarskiptastofa.is/english/telecom-affairs/maritime-communications}
}
\usage{
mmsi_icelandic_registry(con)
}
\arguments{
\item{con}{sql connection}
}
\value{
sql query
}
\description{
Table containing among other things vessel MMSI, including auxillary numbers such as
navigational aid (like net buoy)
and daughter craft. Last version obtained 2022-12-02
}
