% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmsi.R
\name{mmsi_mid}
\alias{mmsi_mid}
\title{Maritime identification digits - country code}
\source{
\url{https://en.wikipedia.org/wiki/Maritime_identification_digits}
}
\usage{
mmsi_mid(con)
}
\arguments{
\item{con}{Oracle connection}
}
\value{
An sql query
}
\description{
Maritime identification digits - country code
}
