% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmsi.R
\name{mmsi_vessel}
\alias{mmsi_vessel}
\title{Vessel MMSI}
\source{
\url{https://www.fjarskiptastofa.is}
}
\usage{
mmsi_vessel(con)
}
\arguments{
\item{con}{sql connection}
}
\value{
sql query
}
\description{
Table containing vessel MMSI, including auxillary numbers such as
navigational aid (like net buoy)
and daughter craft. Last version obtained 2022-04-05
}
