% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Omar-package.R
\docType{package}
\name{Omar-package}
\alias{Omar}
\alias{Omar-package}
\title{Omar: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/omarjcamanto/Omar}
  \item Report bugs at \url{https://github.com/omarjcamanto/Omar/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
