% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pame.R
\name{pame_trail}
\alias{pame_trail}
\title{Connect to pame trail}
\usage{
pame_trail(con, std = TRUE, trim = TRUE)
}
\arguments{
\item{con}{Oracle connection}

\item{std}{Boolean (default TRUE)}

\item{trim}{Boolean (default TRUE)}
}
\value{
A query
}
\description{
Connect to pame trail
}
