% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pame.R
\name{pame_vid_mmsi}
\alias{pame_vid_mmsi}
\title{Pame vessel id}
\usage{
pame_vid_mmsi(con)
}
\arguments{
\item{con}{Oracle connection}
}
\value{
a query
}
\description{
A table of Icelandic vessels id (vid) that are in pame, corresponding mmsi and
the number of records
}
