% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profit_analysis.R
\name{profit_analysis}
\alias{profit_analysis}
\title{Profit analysis}
\usage{
profit_analysis(data, size = 1)
}
\arguments{
\item{data}{input data frame specifying price and acreage of each comparable lot.}

\item{size}{a numerical vector specifying the acreage of lot of interest.}
}
\value{
list
}
\description{
Profit analysis
}
\examples{
df <- dplyr::tribble(
~price, ~acres,
6969, 6,
6996, 9,
9669, 6,
9696, 9
)
profit_analysis(df)
}
