% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sid.R
\name{sid_lwcoeffs}
\alias{sid_lwcoeffs}
\title{biota length-weight coefficients}
\usage{
sid_lwcoeffs(con)
}
\arguments{
\item{con}{oracle connection}
}
\value{
a query
}
\description{
biota length-weight coefficients
}
