% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{smx_len}
\alias{smx_len}
\title{smx_lenth}
\usage{
smx_len(
  con,
  sid = 1,
  sclass = c(19, 30, 31, 34, 35, 37, 40),
  stratification = "new_strata"
)
}
\arguments{
\item{con}{oracle connection}

\item{sid}{species numerical code}

\item{sclass}{A numerical vector indicating sampling class (synaflokkur)}

\item{stratification}{Any or all of "old_strata", "new_strata", "ghl_strata"}
}
\value{
a tibble query
}
\description{
Returns a
}
