% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{smx_st}
\alias{smx_st}
\title{smx_st}
\usage{
smx_st(
  con,
  sclass = c(19, 30, 31, 34, 35, 37, 40),
  stratification = "new_strata",
  trim = TRUE,
  trim_towlength = TRUE
)
}
\arguments{
\item{con}{oracle connection}

\item{sclass}{A numerical vector indicating sampling class (synaflokkur)}

\item{stratification}{Any of "old_strata", "new_strata", "ghl_strata"}

\item{trim}{return only "key" variables (default TRUE)}

\item{trim_towlength}{A boolean (default TRUE) that trims towlength}
}
\value{
a tibble query
}
\description{
Returns a station table by a join of channel.stod_v and channel.syni_v. Generall
indented for standariszed survey.
}
