% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gid.R
\name{std_isscfg}
\alias{std_isscfg}
\title{Title}
\usage{
std_isscfg(con)
}
\arguments{
\item{con}{oracle connection}
}
\value{
A query
}
\description{
International Standard Statistical Classification of Fishing Gear (ISSCFG), see http://www.fao.org/fishery/docs/DOCUMENT/cwp/handbook/annex/AnnexM2fishinggear.pdf
}
