% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vid.R
\name{stk_mapdeck}
\alias{stk_mapdeck}
\title{Title}
\usage{
stk_mapdeck(con, MID)
}
\arguments{
\item{con}{Oracle connection of a data frame}

\item{MID}{mobileid}
}
\value{
q mapdeck object
}
\description{
Title
}
