% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk.R
\name{stk_mobile_icelandic}
\alias{stk_mobile_icelandic}
\title{Link vessel id and mobileid}
\usage{
stk_mobile_icelandic(con, correct = FALSE, vidmatch = FALSE, classify = FALSE)
}
\arguments{
\item{con}{The connection}

\item{correct}{Boolean, if FALSE (default) only returns stk.mobile as is}

\item{vidmatch}{Boolean, if FALSE (default) no matching attempted}

\item{classify}{Boolean, if FALSE (default) no classification attempted}
}
\value{
a query
}
\description{
Link vessel id and mobileid
}
