% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk.R
\name{stk_trail}
\alias{stk_trail}
\title{The AIS/VMS trail}
\usage{
stk_trail(con, year)
}
\arguments{
\item{con}{The connection}

\item{year}{An atomic value (only one year) to narrow time window}
}
\value{
a query
}
\description{
The AIS/VMS trail
}
