% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{tbl_mar}
\alias{tbl_mar}
\title{Connection to MAR tables}
\usage{
tbl_mar(con, tbl)
}
\arguments{
\item{con}{The connection}

\item{tbl}{The name of the table (normally schema.table)}
}
\value{
a query
}
\description{
Connection to MAR tables
}
