% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VESSEL.PAME}
\alias{VESSEL.PAME}
\title{Vessel tracks and behaviour of creel fisheries}
\format{
A data frame with xx rows and yy variables:
\describe{
\item{mmsi2}{vessel identification}
\item{imo2}{record number, unique for each vessel}
\item{n.pings}{time of position}
\item{vessel}{Easting}
\item{flag}{Northing}
\item{astd}{Longitude}
\item{tmin}{Latitude}
\item{tmax}{Vessel activity}}
}
\source{
XXX
}
\usage{
VESSEL.PAME
}
\description{
Vessel tracks and behaviour of creel fisheries
}
\keyword{datasets}
