% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vid2.R
\name{vessel}
\alias{vessel}
\title{The new vessel stuff}
\usage{
vessel(con, fix = TRUE)
}
\arguments{
\item{con}{oracle connection}

\item{fix}{Fix typos, boolean (default TRUE)}
}
\value{
a query
}
\description{
The new vessel stuff
}
