% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vid.R
\name{vessel_cs_itu_prefix}
\alias{vessel_cs_itu_prefix}
\title{Vessel callsign country prefixes}
\source{
\url{https://en.wikipedia.org/wiki/ITU_prefix#Allocation_table}
}
\usage{
vessel_cs_itu_prefix(con)
}
\arguments{
\item{con}{sql connection}
}
\value{
a query
}
\description{
Vessel callsign country prefixes
}
