% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vid.R
\name{vessel_history}
\alias{vessel_history}
\alias{vid_history}
\title{Vessel history per recorded by Fiskistofa}
\usage{
vessel_history(con)

vid_history(con)
}
\arguments{
\item{con}{oracle connection}
}
\value{
a query
}
\description{
Vessel history per recorded by Fiskistofa
}
