% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vid.R
\name{vessel_mclass}
\alias{vessel_mclass}
\title{Vessel management class}
\usage{
vessel_mclass(con, trim = TRUE)
}
\arguments{
\item{con}{oracle connection}

\item{trim}{Boolean (default TRUE) returning only limited variables}
}
\value{
A query
}
\description{
Vessel management class
}
