% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vid.R
\name{vessel_ust}
\alias{vessel_ust}
\title{Icelandic vessel local identification letters}
\usage{
vessel_ust(con)
}
\arguments{
\item{con}{sql connection}
}
\value{
sql query
}
\description{
Table containing local identification letters and location name
}
