% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vid.R
\name{vessels_vessels}
\alias{vessels_vessels}
\title{The mother of all vessels}
\usage{
vessels_vessels(con)
}
\arguments{
\item{con}{oracle connection}
}
\value{
A query
}
\description{
Data from Samgongustofa, Directorates of Fisheries and some from MMSI registry
}
