% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vocabulary}
\alias{vocabulary}
\title{Variable vocabulary}
\format{
\subsection{\code{vocabulary}}{

A vector
}
}
\usage{
vocabulary
}
\description{
Standard variable names
}
\keyword{datasets}
