% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biota.R
\name{bi_len}
\alias{bi_len}
\title{bi_len}
\usage{
bi_len(con, std = TRUE, trim = TRUE, weights = TRUE)
}
\arguments{
\item{con}{oracle connection}

\item{std}{create standardized/shortcut names (default is TRUE)}

\item{trim}{trim return only key variables (default is TRUE). Only operational
if std is TRUE.}

\item{weights}{Gets or calculates weights (default is TRUE). Only operational
if std is TRUE.}
}
\value{
a tibble query with the following variables if arguement std is TRUE (default):
\describe{
\item{.id}{synis_id)}
\item{sid}{species id}
\item{length}{length, normally in centimeters)}
\item{wt}{derived weights in grams, estimated from length weight relationship. If parameters not available in biota.lw_coeffs then 0.01 and 3 are used}
\item{sex}{sex, normally 1 is male, 2 is female}
\item{mat}{maturity stage}
\item{age}{age, normally in years}
\item{n}{number of measured fish}
\item{rn}{the count raising factor}
\item{mwt]{the estimated weights in grams for a given length}
}
}
\description{
Reads biota.lengd_skalad_v that includes a raising factor and
provides weight as a function of length. No raising done.
}
\details{
NOTE: If species only counted at station, not included in algorithm.
}
