% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel.R
\name{ch_station}
\alias{ch_station}
\title{ch_station}
\usage{
ch_station(con, std = TRUE, trim = TRUE)
}
\arguments{
\item{con}{oracle connection}

\item{std}{create standardized/shortcut names (default is TRUE)}

\item{trim}{trim return only key variables (default is TRUE). only operational
if std is TRUE}
}
\value{
a tibble query
}
\description{
Reads channel.stod_v. Access the same view as mar::les_stod
}
