% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Omar-package.R
\docType{package}
\name{Omar-package}
\alias{Omar}
\alias{Omar-package}
\title{Omar: Omar's R Sandbox}
\description{
Generates Rough Guidelines for Optimal Profit.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/omarjcamanto/Omar}
  \item Report bugs at \url{https://github.com/omarjcamanto/Omar/issues}
}

}
\author{
\strong{Maintainer}: Omar Camanto \email{ocamanto@uwo.ca}

}
\keyword{internal}
