% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profit_analysis.R
\name{profit_analysis}
\alias{profit_analysis}
\title{Profit analysis}
\usage{
profit_analysis(data, size = 1)
}
\arguments{
\item{data}{input data frame specifying price and acreage of each comparable lot.}

\item{size}{a numerical vector specifying the acreage of the lot of interest. 1 acre is the default.}
}
\value{
list
}
\description{
Profit analysis
}
\examples{
df <- dplyr::tribble(
~price, ~size, ~unit,
6969, 6, "acre",
6996, 9, "sqft",
9669, 6, "acre",
9696, 9, "sqft"
)
profit_analysis(df)
}
