% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{tbl_mar}
\alias{tbl_mar}
\title{Connection to mar}
\usage{
tbl_mar(con, tbl)
}
\arguments{
\item{con}{oracle connection}

\item{tbl}{a string specifying schema.table}
}
\value{
q query
}
\description{
Connection to mar
}
