% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{tbl_mar}
\alias{tbl_mar}
\title{Connection to mar}
\usage{
tbl_mar(con, tbl)
}
\arguments{
\item{con}{tenging við oraclegrunn}

\item{tbl}{nafn oracle töflu}
}
\value{
sql query and data
}
\description{
Connection to mar
}
