% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vid.R
\name{vessel_registry}
\alias{vessel_registry}
\alias{vid_registry}
\title{The official Icelandic vessel registry}
\usage{
vessel_registry(con, standardize = TRUE, trim = TRUE)

vid_registry(con, standardize = TRUE, trim = TRUE)
}
\arguments{
\item{con}{oracle connection}

\item{standardize}{boolean (default TRUE), returns "standardized" variable names}

\item{trim}{boolean (default TRUE), returns only "standardized" variable names}
}
\value{
a query
}
\description{
Access to table kvoti.skipaskra_siglo
}
