% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vid.R
\name{vessel_valid_imo}
\alias{vessel_valid_imo}
\title{Check if IMO is valid}
\usage{
vessel_valid_imo(x)
}
\arguments{
\item{x}{a numerical vector}
}
\value{
A boolean TRUE/FALSE vector
}
\description{
https://en.wikipedia.org/wiki/IMO_number
}
